Working with Yoko's code. Able to randomly generate tags

+++ trunk/auramap --username K2xL.com/Auramap/src/com/auramap/HappyState.java
@@ -3,8 +3,8 @@ package com.auramap;
 public class HappyState {
 
 	public String stateName;
-	public int xVal;
-	public int yVal;
+	public double xVal;
+	public double yVal;
 	
 	public HappyState() {
